# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "citation.is Public API",
    "version": "1.0.0",
    "description": "The citation.is API provides programmatic access to the verified scientific claims corpus. Ground any AI-generated claim in deterministic, source-backed verdicts.\n\n**Authentication:** All endpoints require a Bearer token in the `Authorization` header. Request an API key at https://citation.is/developers.\n\n**Rate limits:** 60 requests/minute on the free tier, 600 requests/minute on Pro.",
    "contact": {
      "name": "citation.is",
      "url": "https://citation.is/contact"
    },
    "license": {
      "name": "CC BY 4.0",
      "url": "https://creativecommons.org/licenses/by/4.0/"
    }
  },
  "servers": [
    {
      "url": "https://ttruthdesk.claims",
      "description": "Production"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/v1/claim/{id}": {
      "get": {
        "operationId": "getClaimById",
        "summary": "Get a verified claim by ID",
        "description": "Fetch a single verified claim, including its verdict, confidence score, source citations, and full provenance.",
        "tags": ["Corpus"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "The ID of the claim to retrieve",
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": {
            "description": "Claim details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClaimResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid claim ID format",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "401": {
            "description": "Missing or invalid Bearer token",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "404": {
            "description": "Claim not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "API key obtained from https://citation.is/developers"
      }
    },
    "schemas": {
...
      "ClaimResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "claim": { "type": "string" },
          "domain": { "type": "string" },
          "verdict": {
            "type": "object",
            "properties": {
              "status": { "type": "string" },
              "confidenceScore": { "type": "number" },
              "rationale": { "type": "string" },
              "method": { "type": "string" },
              "compositeTruthScore": { "type": "number", "nullable": true },
              "compositeTruthLabel": { "type": "string", "nullable": true }
            }
          },
          "entities": {
            "type": "object",
            "properties": {
              "proteinName": { "type": "string", "nullable": true },
              "pdbId": { "type": "string", "nullable": true },
              "organism": { "type": "string", "nullable": true },
              "ligand": { "type": "string", "nullable": true },
              "experimentalMethod": { "type": "string", "nullable": true }
            }
          },
          "evidence": {
            "type": "object",
            "properties": {
              "sourceDocumentId": { "type": "string" },
              "sourcePassage": { "type": "string", "nullable": true },
              "passageConfidence": { "type": "number", "nullable": true },
              "sourceRefs": { "type": "array", "items": { "type": "object" } },
              "pdbEvidence": { "type": "object", "nullable": true },
              "citations": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "id": { "type": "string" },
                    "type": { "type": "string" },
                    "passageText": { "type": "string", "nullable": true },
                    "confidence": { "type": "number" },
                    "createdAt": { "type": "string", "format": "date-time" }
                  }
                }
              }
            }
          },
          "timestamps": {
            "type": "object",
            "properties": {
              "createdAt": { "type": "string", "format": "date-time" },
              "updatedAt": { "type": "string", "format": "date-time" },
              "verifiedAt": { "type": "string", "format": "date-time" }
            }
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "required": ["error", "code"],
        "properties": {
          "error": { "type": "string" },
          "code": { "type": "string" }
        }
      }
    }
  },
...
}